Cleaned up build system, adjusted exports, added ts type support
