+++ packages/react/package.json
@@ -34,7 +34,13 @@
     "vite": "^4.5.0"
   },
   "description": "React wrappers the Synergy Design System",
-  "exports": {},
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    },
+    "./components/*": "./dist/components/*"
+  },
   "homepage": "https://github.com/SickDesignSystem/synergy/tree/main/packages/react",
   "keywords": [
     "React Web Component Wrapper",
@@ -59,5 +65,5 @@
     "lint": "eslint demo --ext ts,tsx --report-unused-disable-directives --max-warnings 0"
   },
   "type": "module",
-  "version": "0.1.0"
+  "version": "0.0.0"
 }
