--- packages/react/scripts/shared.js
@@ -1,18 +1,0 @@
-export const getAllComponents = metadata => {
-  const allComponents = [];
-
-  metadata.modules.map(module => {
-    module.declarations?.map(declaration => {
-      if (declaration.customElement) {
-        const component = declaration;
-        const path = module.path;
-
-        if (component) {
-          allComponents.push(Object.assign(component, { path }));
-        }
-      }
-    });
-  });
-
-  return allComponents;
-};
