+++ packages/react/src/components/syn-icon/index.ts
