+++ packages/react/src/components/syn-input/index.ts
