+++ packages/react/src/components/syn-spinner/index.ts
