+++ packages/react/src/index.ts
@@ -1,4 +1,4 @@
-export { SynButton } from './syn-button/index.js';
-export { SynIcon } from './syn-icon/index.js';
-export { SynSpinner } from './syn-spinner/index.js';
-export { SynInput } from './syn-input/index.js';
+export { SynButton } from './components/syn-button/index.js';
+export { SynIcon } from './components/syn-icon/index.js';
+export { SynSpinner } from './components/syn-spinner/index.js';
+export { SynInput } from './components/syn-input/index.js';
